requirements.txt
# Core data processing
pandas>=1.5.0
numpy>=1.20.0

# API and data fetching
requests>=2.28.0
python-dotenv>=0.19.0

# Alpaca API (using the newer alpaca-py instead of deprecated alpaca-trade-api)
alpaca-py>=0.8.0

# Database connections
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0

# Data storage formats
pyarrow>=10.0.0  # For Parquet support

# Snowflake connection
snowflake-connector-python>=3.0.0
snowflake-sqlalchemy>=1.4.0

# Data analysis and backtesting
ta-lib>=0.4.0  # Technical analysis indicators
matplotlib>=3.5.0
seaborn>=0.11.0

# Optional: For advanced backtesting
# vectorbt>=0.25.0
# backtrader>=1.9.0

# Jupyter for notebooks
jupyter>=1.0.0
ipython>=8.0.0
